+++ src/index.ts
@@ -1,2 +1,3 @@
 export { apolloExpress, graphiqlExpress } from './integrations/expressApollo';
 export { ApolloHAPI, GraphiQLHAPI } from './integrations/hapiApollo';
+export { apolloKoa } from './integrations/koaApollo';