+++ src/test/tests.ts
@@ -10,4 +10,5 @@ import '../core/runQuery.test';
 import '../modules/operationStore.test';
 import '../integrations/expressApollo.test';
 import '../integrations/hapiApollo.test';
+import '../integrations/koaApollo.test';
 import './testApolloServerHTTP';